requirements.txt
discord.py>=2.0.0
PyYAML>=6.0
pydantic>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.18.0
